feat: add skeleton of flux remote submit
This adds the basic structure for flux remote submit,
which basically wraps the flux submit command (and is
designed to support other subcommands in the future).
This can be moved into the flux libexec cmd directory
or just used as is. I next need to parse the submit
request and then map attributes, etc. to the user
defined compatibility data (cluster and resource needs)
and I think then we are good.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>