feat: support for export.
This allows adding an --export flag to flux detect, which will
either create a temporary context and detect (and export to zip)
all local subsytems, or if given a config directory, run
detect there (honoring force or not) and then export. Of course
not cleaning up if it is not a temporary cluster/subsystem root.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>